+++ tests/bound/nestedwhileloop/gradual_bound/dealloc/nestedwhileloop.c
@@ -3,14 +3,14 @@ int64_t _f_(int64_t limit){
 	int64_t _1 = 0;
 	int64_t i = 0;
 	int64_t sum = 0;
-	int64_t j = 0;
+	uint16_t j = 0;
 	uint16_t _5 = 0;
 	uint16_t _6 = 0;
 	uint16_t _7 = 0;
 	int64_t _8 = 0;
 	int64_t _9 = 0;
 	uint16_t _10 = 0;
-	uint16_t _11 = 0;
+	int64_t _11 = 0;
 	uint16_t _12 = 0;
 	int64_t _13 = 0;
 	//const %5 = 0 : int
